MP-606 :truck: Rename: 테스트 객체를 적절한 곳으로 이동
