+++ src/test/java/kr/modusplant/domains/comment/common/util/domain/event/CommentNotificationEventTestUtils.java
@@ -1,4 +1,4 @@
-package kr.modusplant.shared.event.common.util;
+package kr.modusplant.domains.comment.common.util.domain.event;
 
 import kr.modusplant.domains.comment.domain.event.CommentRegisterEvent;
