+++ src/test/java/kr/modusplant/domains/member/domain/event/CommentLikeEventTest.java
@@ -1,4 +1,4 @@
-package kr.modusplant.shared.event;
+package kr.modusplant.domains.member.domain.event;
 
 import kr.modusplant.domains.member.domain.event.CommentLikeEvent;
 import kr.modusplant.shared.exception.InvalidValueException;
