+++ src/test/java/kr/modusplant/domains/notification/adapter/controller/NotificationControllerTest.java
@@ -11,9 +11,9 @@ import kr.modusplant.domains.notification.usecase.record.NotificationPreview;
 import kr.modusplant.domains.notification.usecase.record.NotificationReadModel;
 import kr.modusplant.domains.notification.usecase.response.CursorPageResponse;
 import kr.modusplant.domains.notification.usecase.response.NotificationResponse;
-import kr.modusplant.shared.event.common.util.CommentLikeNotificationEventTestUtils;
-import kr.modusplant.shared.event.common.util.CommentNotificationEventTestUtils;
-import kr.modusplant.shared.event.common.util.PostLikeNotificationEventTestUtils;
+import kr.modusplant.domains.member.common.util.domain.event.CommentLikeNotificationEventTestUtils;
+import kr.modusplant.domains.comment.common.util.domain.event.CommentNotificationEventTestUtils;
+import kr.modusplant.domains.member.common.util.domain.event.PostLikeNotificationEventTestUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
